Work with float instead of double, to save memory for big catalogues. Currently rejects numpy arrays of type double!

+++ src/band_spec.c
@@ -5,4 +5,4 @@ Data for band specifications.
 #include "ghmock.h"
 
 const char BAND_NAMES[NUM_BANDS] = "ugriz";
-const double BAND_WAVELENGTHS[NUM_BANDS] = {3543., 4770., 6231., 7625., 9134.};
+const float BAND_WAVELENGTHS[NUM_BANDS] = {3543., 4770., 6231., 7625., 9134.};
